Implement paginated logs for all storage backends
