+++ common/src/main/java/me/lucko/luckperms/common/storage/implementation/file/FileActionLogger.java
@@ -25,12 +25,14 @@
 
 package me.lucko.luckperms.common.storage.implementation.file;
 
+import com.google.common.collect.ImmutableList;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.stream.JsonReader;
 import me.lucko.luckperms.common.actionlog.ActionJsonSerializer;
 import me.lucko.luckperms.common.actionlog.Log;
 import me.lucko.luckperms.common.actionlog.LogPage;
+import me.lucko.luckperms.common.actionlog.LoggedAction;
 import me.lucko.luckperms.common.cache.BufferedRequest;
 import me.lucko.luckperms.common.filter.FilterList;
 import me.lucko.luckperms.common.filter.PageParameters;
@@ -45,11 +47,14 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.ReentrantLock;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class FileActionLogger {
 
@@ -133,36 +138,37 @@ public class FileActionLogger {
         }
     }
 
-    public Log getLog() throws IOException {
+    private Stream<LoggedAction> getRawLog() throws IOException {
         // if there is log content waiting to be written, flush immediately before trying to read
         if (this.saveBuffer.isEnqueued()) {
             this.saveBuffer.requestDirectly();
         }
 
         if (!Files.exists(this.contentFile)) {
-            return Log.empty();
+            return Stream.empty();
         }
 
-        Log.Builder log = Log.builder();
-
+        Stream.Builder<LoggedAction> builder = Stream.builder();
         try (BufferedReader reader = Files.newBufferedReader(this.contentFile, StandardCharsets.UTF_8)) {
             String line;
             while ((line = reader.readLine()) != null) {
                 try {
                     JsonElement parsed = GsonProvider.parser().parse(line);
-                    log.add(ActionJsonSerializer.deserialize(parsed));
+                    builder.add(ActionJsonSerializer.deserialize(parsed));
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
             }
         }
+        return builder.build();
+    }
 
-        return log.build();
+    public Log getLog() throws IOException {
+        return Log.of(getRawLog().collect(Collectors.toList()));
     }
 
     public LogPage getLogPage(FilterList<Action> filters, PageParameters page) throws IOException {
-
-
+        return LogPage.of(page.paginate(getRawLog().filter(filters::evaluate).sorted(Comparator.reverseOrder())).collect(Collectors.toList()));
     }
 
     private final class SaveBuffer extends BufferedRequest<Void> {
